Major refactor separating web and simulator code for multithreading
