+++ src/networkapi.h
@@ -4,13 +4,12 @@
 namespace argos {
   class CNetworkAPI;
 }
-
 #include <argos3/core/simulator/visualization/visualization.h>
 #include <sys/time.h>
-#include <nlohmann/json.hpp>
-#include <thread>
-#include "App.h"
+#include <loguru.hpp>
 #include "helpers/Timer.h"
+
+#include "networkapi_webserver.h"
 
 namespace argos {
 
@@ -60,21 +59,6 @@ namespace argos {
     void ResetExperiment();
 
    private:
-    /** Broadcasts a message to all the connected clients */
-    void BroadcastMessage(std::string_view message);
-
-    /** Broadcasts on event channel to all the connected clients */
-    void EmitEvent(std::string_view event_name);
-
-    /** Broadcasts game state to all the connected clients */
-    void BroadcastState();
-
-    /** Function to setup all routes and webhooks */
-    void SetupWebApp(uWS::App& c_MyApp);
-
-    /** Function to send JSON over HttpResponse */
-    void SendJSON(uWS::HttpResponse<false>* pc_res, nlohmann::json c_json_data);
-
     /** Function to run simulation step in realtime */
     void RealTimeStep();
 
@@ -87,40 +71,10 @@ namespace argos {
       EXPERIMENT_SUSPENDED,
       EXPERIMENT_DONE
     };
-    inline std::string EExperimentStateToString(EExperimentState type) {
-      switch (type) {
-        case EXPERIMENT_INITIALIZED:
-          return "EXPERIMENT_INITIALIZED";
-        case EXPERIMENT_PLAYING:
-          return "EXPERIMENT_PLAYING";
-        case EXPERIMENT_FAST_FORWARDING:
-          return "EXPERIMENT_FAST_FORWARDING";
-        case EXPERIMENT_PAUSED:
-          return "EXPERIMENT_PAUSED";
-        case EXPERIMENT_SUSPENDED:
-          return "EXPERIMENT_SUSPENDED";
-        case EXPERIMENT_DONE:
-          return "EXPERIMENT_DONE";
-        default:
-          return "unknown";
-      }
-    }
 
    private:
-    /** HTTP Port to Listen to */
-    UInt16 m_unPort;
-
-    /* Data attached to each socket,
-     * ws->getUserData returns one of these */
-    struct m_sPerSocketData {};
-
     /* Experiment State */
     EExperimentState m_eExperimentState;
-
-    /** Threads serving web requests */
-    std::vector<uWS::WebSocket<false, true>*> m_vecWebSocketClients;
-    /** List of all WebSocket clients connected */
-    std::vector<std::thread*> m_vecWebThreads;
 
     /** True when fast forwarding */
     bool m_bFastForwarding;
@@ -130,6 +84,9 @@ namespace argos {
 
     /** Milliseconds required for one tick of simulator */
     std::chrono::milliseconds m_cSimulatorTickMillis;
+
+    /** Webserver */
+    NetworkAPI::CWebServer m_cWebServer;
   };
 
 }  // namespace argos
